frequency was supposed to be an integer, not a float

+++ db/migrate/20190702010935_create_word_frequencies.rb
@@ -3,7 +3,7 @@ class CreateWordFrequencies < ActiveRecord::Migration[5.2]
     create_table :word_frequencies do |t|
       t.string :name
       t.integer :rank
-      t.float :frequency
+      t.integer :frequency
 
       t.timestamps
     end
